+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+- Feature: Limit amount of transactions in a single database call with `max_db_batch_size`
+
 # 1.19.0 - 2023-02-21
 
 - Feature: When consuming data in batch mode, this adds ability to save data to multiple tables with `association_list`.